Done Sprint-4-task-2 CR#2
Fixed all the comments

+++ db/schema.rb
@@ -43,7 +43,7 @@ ActiveRecord::Schema.define(version: 2023_03_09_082142) do
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
     t.integer "number_of_available_kudos", default: 10, null: false
-    t.integer "number_of_earned_points", default: 0
+    t.float "number_of_earned_points", default: 0.0
     t.index ["email"], name: "index_employees_on_email", unique: true
     t.index ["reset_password_token"], name: "index_employees_on_reset_password_token", unique: true
   end
